requirements.txt
PyQt5==5.11.3
pyserial==3.4
